Adding file upload

+++ app/Http/Controllers/PostsController.php
@@ -3,13 +3,13 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
-
+use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Facades\Image;
 use App\Models\Post;
+use DB;
 
 class PostsController extends Controller
 {
-
     /**
      * Create a new controller instance.
      *
@@ -17,9 +17,9 @@ class PostsController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('auth', ['except' => ['index','show']]);
+        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
-    
+
     /**
      * Display a listing of the resource.
      *
@@ -27,8 +27,14 @@ class PostsController extends Controller
      */
     public function index()
     {
+        //$posts = Post::all();
+        //return Post::where('title', 'Post Two')->get();
+        //$posts = DB::select('SELECT * FROM posts');
+        //$posts = Post::orderBy('title','desc')->take(1)->get();
+        //$posts = Post::orderBy('title','desc')->get();
+
         $posts = Post::orderBy('created_at','desc')->paginate(10);
-        return view('posts.index')->with('posts',$posts);
+        return view('posts.index')->with('posts', $posts);
     }
 
     /**
@@ -51,20 +57,42 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+		
+	    // make thumbnails
+	    $thumbStore = 'thumb.'.$filename.'_'.time().'.'.$extension;
+            $thumb = Image::make($request->file('cover_image')->getRealPath());
+            $thumb->resize(80, 80);
+            $thumb->save('storage/cover_images/'.$thumbStore);
+		
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
-        //Create A Post
+        // Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
-        
+
         return redirect('/posts')->with('success', 'Post Created');
-
-
     }
 
     /**
@@ -76,7 +104,7 @@ class PostsController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
-        return view('posts.show')->with('post',$post);
+        return view('posts.show')->with('post', $post);
     }
 
     /**
@@ -88,15 +116,18 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
-        //
-        if(auth()->user()->id !==$post->user_id){
-            return redirect('/posts')->with('error', 'Unauthorised Page');
+        
+        //Check if post exists before deleting
+        if (!isset($post)){
+            return redirect('/posts')->with('error', 'No Post Found');
         }
 
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
 
-
-
-        return view('posts.edit')->with('post',$post);
+        return view('posts.edit')->with('post', $post);
     }
 
     /**
@@ -112,15 +143,38 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
+		$post = Post::find($id);
+         // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            // Delete file if exists
+            Storage::delete('public/cover_images/'.$post->cover_image);
+		
+	   //Make thumbnails
+	    $thumbStore = 'thumb.'.$filename.'_'.time().'.'.$extension;
+            $thumb = Image::make($request->file('cover_image')->getRealPath());
+            $thumb->resize(80, 80);
+            $thumb->save('storage/cover_images/'.$thumbStore);
+		
+        }
 
-
-        //Create A Post
-        $post = Post::find($id);
+        // Update Post
         $post->title = $request->input('title');
         $post->body = $request->input('body');
-        $post->user_id = auth()->user()->id;
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
-        
+
         return redirect('/posts')->with('success', 'Post Updated');
     }
 
@@ -132,12 +186,24 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-       $post = Post::find($id);
+        $post = Post::find($id);
+        
+        //Check if post exists before deleting
+        if (!isset($post)){
+            return redirect('/posts')->with('error', 'No Post Found');
+        }
 
-       if(auth()->user()->id !==$post->user_id){
-        return redirect('/posts')->with('error', 'Unauthorised Page');
-    }
-       $post->delete();
-       return redirect('/posts')->with('success','Post Removed');
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+        
+        $post->delete();
+        return redirect('/posts')->with('success', 'Post Removed');
     }
 }
